[Scripts] Allow run scripts without waiting for the launcher process to end -> Refactor platform into Utils::Platform then create ProcessStartInfo class
